backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
sentence-transformers
faiss-cpu
pypdf
python-dotenv
langchain
langchain-community
langchain-google-genai
numpy
scikit-learn
pydantic
requests
markdown
beautifulsoup4
chardet
PyPDF2
tqdm
typing-extensions
urllib3
openai
tiktoken
huggingface-hub
accelerate
safetensors
tokenizers
chromadb
pytest
httpx
pytest-asyncio
black
isort
flake8
mypy
pre-commit
poetry
pip-tools
build
twine
wheel
setuptools
virtualenv
pip
python-docx
